third commit

+++ RISCV_ISA_SIM/decode.hpp
@@ -27,7 +27,7 @@
 /*  position:  31...       y  ...   x ...  210  */
 /*  for all k if (31>=x>=k>=y>=0), bit(k) = 1,  */
 /*                      otherwise, bit(k) = 0   */
-#define ONES(x,y)       (reg32) (((1<<(x+1))-1)-((1<<y)-1))
+#define ONES(x,y)       (reg32) ((((long)1<<(x+1))-1)-((1<<y)-1))
 
 /* masks */                                          // bit LEN
 #define OPCODE     ONES(6,0)      // 7
